% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ebay_token}
\alias{get_ebay_token}
\title{Get eBay Token}
\usage{
get_ebay_token()
}
\value{
a character string, containing the eBay API token.
}
\description{
This function returns the eBay api token stored as an environmental variable.
}
