% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_ebay}
\alias{search_ebay}
\title{Search eBay Listings}
\usage{
search_ebay(keywords, type = c("active", "completed"),
  search_description = TRUE, categories = NULL, item_filters = NULL,
  n_results = 100, token = get_ebay_token())
}
\arguments{
\item{keywords}{a character string, containing keywords for ebay listings search}

\item{type}{a character string, either "active" or "completed" to indicate
the type of listings to search.  Defaults to "active".}

\item{search_description}{logical, set to TRUE to include sub-titles and
item descriptions in the search. If FALSE, the search will only look
at titles of items.}

\item{categories}{character (or numeric) vector (max length = 3), containing
the categories to include in search.}

\item{item_filters}{a named list, containing and adhering to the conventions
specified in the eBay API documention.
\href{https://developer.ebay.com/Devzone/finding/CallRef/types/ItemFilterType.html}{ItemFilter Documentation}}

\item{n_results}{numeric (length 1), the number of listing results to return.
Rounded to the nearest 100.}

\item{token}{eBay Devleoper API token (AppID)}
}
\value{
dataframe, containing eBay search results
}
\description{
Search eBay Listings
}
\examples{
df.items <- search_ebay(
  "model train",
  type = "active",
  item_filters = list(MinPrice = 100, MaxPrice = 300)
)
}
