% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ebay_token}
\alias{set_ebay_token}
\title{Set eBay Token}
\usage{
set_ebay_token(token)
}
\arguments{
\item{token}{a character string, containing eBay API token.}
}
\description{
This function allows you to set and store your eBay API token as an
    environmental variable.
}
